Match filetype instead of equating
Instead of comparing the full filetype against 'markdown', check if
the latter is a subtring of the former.

The rationale for this is that [vimwiki](https://vimwiki.github.io)
(and perhaps some other plugins) concatenate filetype names. For
example, if `let g:vimwiki_filetypes = ['markdown']`, then the
filetype for vimwiki files is `vimwiki.markdown`.