feat(#20): percent-encode paths
The primary intention is to fix the case where filenames have spaces in
them and this plugin doesn't encode the URL properly so it appears as if
vivify can't refresh

We do this whether (n)vim has python3 support or not, but with python3
support do a proper encoding with a library to handle more esoteric
cases, for example curly brackets in filenames.

We found that even without the library, characters such as ä, á, ñ, €
already work, so it's probably a very rare case where you need any more
meticulous encoding for filenames. And then even if that occurs,
enabling python3 support will solve it :)